Transformer/requirements.txt
anyio==4.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
blas==1.0
bleach==6.2.0
bottleneck==1.4.2
brotli-python==1.0.9
bzip2==1.0.8
ca-certificates==2024.12.31
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.3.1
cpuonly==2.0
cycler==0.11.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.0
executing==0.8.3
filelock==3.13.1
fonttools==4.55.3
freetype==2.12.1
giflib==5.2.2
gmpy2==2.1.2
h11==0.14.0
httpcore==1.0.2
httpx==0.27.0
icu==73.1
idna==3.7
intel-openmp==2023.1.0
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
jinja2==3.1.4
jpeg==9e
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.7.2
jupyter_events==0.10.0
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.2.5
jupyterlab_pygments==0.1.2
jupyterlab_server==2.27.3
kiwisolver==1.4.8
krb5==1.20.1
lcms2==2.16
lerc==4.0.0
libabseil==20240116.2
libclang==14.0.6
libclang13==14.0.6
libdeflate==1.22
libffi==3.4.4
libjpeg-turbo==2.0.0
libpng==1.6.39
libpq==17.2
libprotobuf==4.25.3
libsodium==1.0.18
libtiff==4.5.1
libuv==1.48.0
libwebp==1.3.2
libwebp-base==1.3.2
lz4-c==1.9.4
markupsafe==2.1.3
matplotlib==3.10.0
matplotlib-base==3.10.0
matplotlib-inline==0.1.6
mistune==2.0.4
mkl==2023.1.0
mkl-service==2.4.0
mkl_fft==1.3.11
mkl_random==1.2.8
mpc==1.1.0
mpfr==4.0.2
mpir==3.0.0
mpmath==1.3.0
mysql-connector-python==8.4.0
nbclient==0.8.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.2
notebook-shim==0.2.3
numexpr==2.10.1
numpy==1.25.2
numpy-base==1.25.2
openjpeg==2.5.2
openssl==3.0.15
overrides==7.4.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.4
pillow==11.0.0
pip==24.2
platformdirs==3.10.0
ply==3.11
prometheus_client==0.21.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.0
pure_eval==0.2.2
pycparser==2.21
pygments==2.15.1
pyparsing==3.2.0
pyqt==5.15.10
pysocks==1.7.1
python==3.10.16
python-dateutil==2.9.0post0
python-dotenv==0.21.0
python-fastjsonschema==2.20.0
python-json-logger==3.2.1
pytorch==2.5.1
pywin32==308
pywinpty==2.0.14
pyyaml==6.0.2
pyzmq==26.2.0
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
sqlite==3.45.3
stack_data==0.2.0
sympy==1.13.3
tbb==2021.8.0
terminado==0.17.1
tinycss2==1.2.1
tk==8.6.14
tomli==2.0.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2024b
unicodedata2==15.1.0
urllib3==2.3.0
vc==14.40
vs2015_runtime==14.42.34433
wcwidth==0.2.5
websocket-client==1.8.0
wheel==0.44.0
xz==5.4.6
yaml==0.2.5
zlib==1.2.13
zstd==1.5.6
